+++ eslint.config.mjs
@@ -1,0 +1,12 @@
+import { recommended } from '@nextcloud/eslint-config'
+
+export default [
+	...recommended,
+	{
+		ignores: [
+			'js/**',
+			'node_modules/**',
+			'vendor/**',
+		],
+	},
+]