+++ src/adminSettings.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
-import { applyGlobals } from './bootstrap.js'
 import AdminSettings from './components/AdminSettings.vue'
+import { applyGlobals } from './bootstrap.js'
 
 const app = createApp(AdminSettings)
 applyGlobals(app)