+++ src/personalSettings.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
-import { applyGlobals } from './bootstrap.js'
 import PersonalSettings from './components/PersonalSettings.vue'
+import { applyGlobals } from './bootstrap.js'
 
 const app = createApp(PersonalSettings)
 applyGlobals(app)